[#69] Fix action creation
Action creation was completely broken. Variables were referenced before being
declared, messages were written to the console before it was in scope, and the
route was using a "controller" segment, which will not work, as it overwrites
the default for the controller to dispatch. Paths for view scripts were being
generated incorrectly, and code reflection/generation were being done in a way
that didn't work. This commit fixes all of that.